refactor (api-objects): change file paths - export all public modules of basic message components & search response formats from /api-objects
